backend/requirements.txt
google-genai
qdrant-client
fastapi
uvicorn[standard]
python-dotenv
pillow
httpx
pydantic
python-multipart
